Injectors refactor works, not too tested

+++ src/stak_func/stak/z_injectors.py
@@ -4,74 +4,109 @@ So, the solution is to compute these things in this file & inject the results in
 Also, removes some clutter from the code.
 Also, settings are injected by running this.
 
-Since blocks are numbered it is very tedious to insert or delete a new block,
-running this file will rename the appropriate files when one of them is deleted
-& insert & rename when "newBlockName" is provided automatically.
+Since blocks are numbered it is very tedious to insert or delete a new block, yes it is essential that they be numbered.
+running this file will rename the appropriate files when one of them is deleted & insert & rename when "newBlockName" is
+provided & more. If this file could rename itself, then it cannot be numbered itself.
 """
+
 import sys
 import os
 
 from collections import defaultdict
 from itertools   import chain, izip
 
-from .block00_typing        import *
+from .block00_typing         import *
 from . import block01_settings   as settings
 from . import block02_settingObj as settingObj
 from . import block03_constants  as cs
-from .block16_utils          import read, write
+from .block15_utils          import read, write
 
-## Block prefix
-newBlockPrefix    = 'block'
-oldBlockPrefix    = cs.blockPrefix
-lenNewBlockPrefix = len(newBlockPrefix)
-lenOldBlockPrefix = len(oldBlockPrefix)
 
 osPath   = os.path
 abspath  = osPath.abspath
 joinPath = osPath.join
+dirname  = osPath.dirname
 
-## Old & new paths
-dirname     = osPath.dirname
-packagePath = dirname(abspath(__file__))
-constsPath  = abspath(cs.__file__)
-
-oldPathSplitChar = '/' if '/' in sys._getframe(0).f_code.co_filename else '\\'
-newPathSplitChar = cs.pathSplitChar
-
-pyExt     = cs.pyExt
-nLenPyExt = -cs.lenPyExt
+pyExt          =  cs.pyExt
+nLenPyExt      = -cs.lenPyExt
+pycExt         =  cs.pycExt
+nLenPycExt     = -cs.lenPycExt
+blockPrefix    =  cs.blockPrefix
+lenBlockPrefix =  cs.lenBlockPrefix
 
 def replaceInPlace(lines, lookingFor, dataStructure):  # type: (Lst[str], str, Any) -> None
     for i, line in enumerate(lines):
         if line.startswith(lookingFor):
-            lines[i] = '%s%r' % (lookingFor, dataStructure)
+            lines[i] = '%s%r\n  # Injected by block15_injectors.py' % (lookingFor, dataStructure)
             return
     print '[STAK] ERROR: Looking for', lookingFor, 'prefix & not found'
 
 
+# Read lazy, written on exit.
+linesByPath = {}  # type: Dic[str, Lst[str]]
+
+def getLinesForPath(_path, _linesByPath=linesByPath):  # type: (str, ...) -> Lst[str]
+    if _path not in _linesByPath:
+        _linesByPath[_path] = read(_path)
+    return _linesByPath[_path]
+
+linesByName = {}
+pathsByName = {}
+def getLinesForName(name):
+    if name not in linesByName:
+        path = pathsByName[name]
+        linesByName[name] = read(path)
+    return linesByName[name]
+
+
+# Parse dir files
+packagePath = dirname(abspath(__file__))
+oldStakFiles = os.listdir(packagePath)
+oldPyFileNames = [name[:nLenPyExt] for name in oldStakFiles if name[nLenPyExt:] == pyExt]
+oldBlockFileNames = [name for name in oldPyFileNames if name[:lenBlockPrefix] == blockPrefix]
+
+# Make old block files names and indexes
+# ---------------------------------------------------------------------------------------------------------------------
+oldBlockFiles   = []; oldBlockFilesApp   = oldBlockFiles.append
+nonBlockFiles   = []; nonBlockFilesApp   = nonBlockFiles.append
+oldBlockNames   = []; oldBlockNamesApp   = oldBlockNames.append
+oldBlockStrIdxs = []; oldBlockStrIdxsApp = oldBlockStrIdxs.append
+
+for file in os.listdir(packagePath):
+    if file[nLenPyExt:] != pyExt:
+        continue
+    file = file[:nLenPyExt]
+
+    if file[:lenBlockPrefix] != blockPrefix:
+        nonBlockFilesApp(file)
+        pathsByName[file] = joinPath(packagePath, file + pyExt)
+        continue
+
+    oldBlockFilesApp(file)
+    file = file[lenBlockPrefix:]
+    idx, name = file.split('_')
+    oldBlockStrIdxsApp(idx)
+    oldBlockNamesApp(name)
+# ---------------------------------------------------------------------------------------------------------------------
+
+## Make newBlockFiles
+# ---------------------------------------------------------------------------------------------------------------------
+newBlockNames = cs.blockNames
+oldMaxBlockIdxLen = len(str(len(oldBlockNames) - 1))
+newMaxBlockIdxLen = len(str(len(newBlockNames) - 1))
+newBlockFiles = [
+    blockPrefix + str(i).zfill(newMaxBlockIdxLen) + '_' + name
+    for i, name in enumerate(newBlockNames)
+]
+# ---------------------------------------------------------------------------------------------------------------------
+
+
 ## On blocks changed run block injection
 # ---------------------------------------------------------------------------------------------------------------------
-listDir = os.listdir
-oldBlockFiles = [
-    name[:nLenPyExt]
-    for name in listDir(packagePath)
-    if name[nLenPyExt:] == pyExt and name != '__init__.py'
-]
-
-oldBlockNames = [name.split('_')[-1] for name in oldBlockFiles]
-newBlockNames = cs.blockNames
-
-blockPrefix = cs.blockPrefix
-
-if oldBlockNames != newBlockNames:
-    oldMaxBlockIdxLen = len(str(len(oldBlockNames) - 1))
-    newMaxBlockIdxLen = len(str(len(newBlockNames) - 1))
-
-    newBlockFiles = [
-        blockPrefix + str(i).zfill(newMaxBlockIdxLen) + '_' + name
-        for i, name in enumerate(newBlockNames)
-    ]
-
+if oldBlockFiles == newBlockFiles:
+    for file, name in izip(oldBlockFiles, oldBlockNames):
+        pathsByName[name] = joinPath(packagePath, name + pyExt)
+else:
     ## Add / remove block names
     # -----------------------------------------------------------------------------------------------------------------
     oldBlockNamesSet = set(oldBlockNames)
@@ -85,23 +120,26 @@ if oldBlockNames != newBlockNames:
         sys.exit()
 
     removeFile = os.remove
-
     for name in blockNamesToRemove:
         fileName = oldBlockFiles[oldBlockNames.index(name)]
+
+        if not input('DANGER sure of removing: %s ??' % fileName):
+            sys.exit()
+
         removeFile(joinPath(packagePath, fileName + pyExt))
 
     for name in blockNamesToAdd:
         fileName = newBlockFiles[newBlockNames.index(name)]
-        write(
-            joinPath(packagePath, fileName + pyExt),
-            ('from .block00_typing import *\n',)
-        )
+        path = joinPath(packagePath, fileName + pyExt)
+        pathsByName[name] = path
+        write(path, ('from .block00_typing import *\n',))
     # -----------------------------------------------------------------------------------------------------------------
 
     ## Make new by old blocks. For renaming in files & therefore excludes new blocks.
     # -----------------------------------------------------------------------------------------------------------------
     blockPrefix = cs.blockPrefix
-    newByOldBlocks = []; append = newByOldBlocks.append
+    newByOldBlocks = []
+    newByOldBlocksApp = newByOldBlocks.append
     enumOldBlockNames = list(enumerate(oldBlockNames))
 
     for newIdx, newName in enumerate(newBlockNames):
@@ -111,13 +149,10 @@ if oldBlockNames != newBlockNames:
         else:
             continue
 
-        oldIdxStr = str(oldIdx).zfill(oldMaxBlockIdxLen)
-        newIdxStr = str(newIdx).zfill(newMaxBlockIdxLen)
-
-        append(
+        newByOldBlocksApp(
             (
-                blockPrefix + oldIdxStr + '_' + oldName,
-                blockPrefix + newIdxStr + '_' + newName,
+                oldBlockFiles[oldIdx],
+                blockPrefix + str(newIdx).zfill(newMaxBlockIdxLen) + '_' + newName,
             )
         )
     # -----------------------------------------------------------------------------------------------------------------
@@ -125,70 +160,71 @@ if oldBlockNames != newBlockNames:
     ## Rename block files.
     # -----------------------------------------------------------------------------------------------------------------
     renamePath = os.rename
-    blockPaths = []; append = blockPaths.append
     for oldBlock, newBlock in newByOldBlocks:
+        name = newBlockNames[newBlockFiles.index(newBlock)]
         newBlockPath = joinPath(packagePath, newBlock + pyExt)
-        append(newBlockPath)
+        pathsByName[name] = newBlockPath
+
         if oldBlock != newBlock:
             oldBlockPath = joinPath(packagePath, oldBlock + pyExt)
             renamePath(oldBlockPath, newBlockPath)
     # -----------------------------------------------------------------------------------------------------------------
 
-    print 'blockPaths'
-    for path in blockPaths:
-        print path
-    sys.exit()
-
     ## Rename references to blocks, sortBlock imports.
     # -----------------------------------------------------------------------------------------------------------------
-    for path in blockPaths:
-        lines = read(path)
+    for name in chain(newBlockNames, nonBlockFiles):
+        lines = getLinesForName(name)
 
         importsStarted = importsFinished = False
-        imports = []; appImports = imports.append
-        lineNos = []; appLineNos = lineNos.append
+        import_blockAndLine = []; import_blockAndLineApp = import_blockAndLine.append
+        lineNos = []; lineNosApp = lineNos.append
 
         for i, line in enumerate(lines):
 
-            anyBlockInLine = False
+            # Rename references to blocks
             for oldBlock, newBlock in newByOldBlocks:
                 if oldBlock in line:
                     line = line.replace(oldBlock, newBlock)
                     lines[i] = line
-                    anyBlockInLine = True
 
+            # Sort imports, only block imports which should be all together by codestyle.
             if importsFinished:
                 continue
 
-            if anyBlockInLine and ' import ' in line:
-                importsStarted = True
-                appImports(line)
-                appLineNos(i)
+            if 'import ' in line:
+                for block in newBlockFiles:
+                    if block in line:
+                        importsStarted = True
+                        import_blockAndLineApp((block, line))
+                        lineNosApp(i)
+                        break
 
             elif importsStarted:
                 importsFinished = True
 
         if importsStarted and importsFinished:
-            imports.sort()
-            for num, _import in izip(lineNos, imports):
-                lines[num] = _import
+
+            import_blockAndLine.sort()
+
+            for (block, line), lineNo in izip(import_blockAndLine, lineNos):
+                lines[lineNo] = line
+
         elif not importsStarted and not importsFinished:
             continue
         else:
             print '[STAK] ERROR: Imports sorting. started', importsStarted, 'finished', importsFinished
 # ---------------------------------------------------------------------------------------------------------------------
 
-
 ## Run manual injectors.
 # ---------------------------------------------------------------------------------------------------------------------
 if cs.runInjectors:
+    def blockNameFromModule(module):  # type: (ModuleType) -> str
+        return module.__name__.split('_')[-1]
+
     # Settings injection
     # -----------------------------------------------------------------------------------------------------------------
-    settingsPath   = settings.__file__
-    settingObjPath = settingObj.__file__
-
-    settingsLines   = read(settingsPath)
-    settingObjLines = read(settingObjPath)
+    settingsLines   = getLinesForName(blockNameFromModule(settings))
+    settingObjLines = getLinesForPath(blockNameFromModule(settingObj))
 
     settingsNames = [name for name in settings.__dict__ if name[-2:] != '__']
 
@@ -222,8 +258,6 @@ if cs.runInjectors:
 
         elif objLine == '    def __init__(self):\n':
             started = 1
-
-    write(settingObjPath, settingObjLines)
     # -----------------------------------------------------------------------------------------------------------------
 
     ## Constants injection
@@ -241,7 +275,7 @@ if cs.runInjectors:
 
     allFlags = list(chain(cs.stakFlags, cs.stdFlags, (cs.cutoffFlag,)))
     pAllFlags = padFlags(allFlags)
-    padByNotAllFlags =  {flag: pFlag for flag, pFlag in izip(allFlags, pAllFlags)}
+    padByNotAllFlags = {flag: pFlag for flag, pFlag in izip(allFlags, pAllFlags)}
 
     ## Make whole enough std flags
     repeatsByCutoffFlags = defaultdict(int)  # type: Dic[str, int]
@@ -261,7 +295,7 @@ if cs.runInjectors:
     ]
 
     ## Inject consts.
-    constsLines = read(constsPath)
+    constsLines = getLinesForPath(blockNameFromModule(cs))
 
     replaceInPlace(constsLines, 'paddedStdFlags = '     , paddedStdFlags)
     replaceInPlace(constsLines, 'pStakFlags = '         , paddedStakFlags)
@@ -270,16 +304,22 @@ if cs.runInjectors:
     replaceInPlace(constsLines, 'allPflagsByFlags = '   , padByNotAllFlags)
     replaceInPlace(constsLines, 'wholeEnoughs = '       , wholeEnoughFlags)
     replaceInPlace(constsLines, 'runInjectors = '       , 0)
-
-    write(constsPath, constsLines)
     # -----------------------------------------------------------------------------------------------------------------
+# ---------------------------------------------------------------------------------------------------------------------
+
+
+## Write all lines.
+# ---------------------------------------------------------------------------------------------------------------------
+for name, lines in linesByName.iteritems():
+    path = pathsByName[name]
+    write(path, lines)
 # ---------------------------------------------------------------------------------------------------------------------
 
 
 ## Unload all blocks
 # ---------------------------------------------------------------------------------------------------------------------
 modules = sys.modules
-packageDotPathPrefix = __name__.split('.')[:-1].join('.') + '.'
+packageDotPathPrefix = '.'.join(__name__.split('.')[:-1]) + '.'
 for name in newBlockNames:
     moduleDotPath = packageDotPathPrefix + name
     if moduleDotPath in modules:
